feat!: move init config & add cache init
